feat(Makefile): add run.artifact target to execute Java application
refactor(cli.py): use OpenAPIConverter directly for spec conversion

refactor(generator.py): remove convert_spec_version method

test(cli.py): update tests to mock OpenAPIConverter instead of OpenAPIGenerator
